Primera iteracion de menu
