+++ src/main/java/org/example/Main.java
@@ -1,19 +1,11 @@
 package org.example;
 
-import org.example.menues.HotelDelSurLanding;
+import org.example.menues.HiloMenues;
 
 import javax.swing.*;
 
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
 public class Main {
     public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            public void run() {
-                HotelDelSurLanding frame = new HotelDelSurLanding();
-                frame.setSize(800, 600);
-                frame.setVisible(true);
-            }
-        });
+        SwingUtilities.invokeLater(new HiloMenues());
     }
 }
