--- src/main/java/org/example/entidades/Reserva.java
@@ -1,64 +1,0 @@
-package org.example.entidades;
-
-import org.example.entidades.persona.Cliente;
-import org.example.enums.Estado;
-
-import java.time.LocalDate;
-
-public class Reserva {
-    private final String idReserva;
-    private final Cliente cliente;
-    private Habitacion habitacion;
-    private LocalDate fechaInicio;
-    private LocalDate fechaFin;
-    private Estado estado;
-
-    public Reserva(String idReserva, Cliente cliente, Habitacion habitacion, LocalDate fechaInicio, LocalDate fechaFin, Estado estado) {
-        this.idReserva = idReserva;
-        this.cliente = cliente;
-        this.habitacion = habitacion;
-        this.fechaInicio = fechaInicio;
-        this.fechaFin = fechaFin;
-        this.estado = estado;
-    }
-
-    public String getIdReserva() {
-        return idReserva;
-    }
-
-    public Cliente getCliente() {
-        return cliente;
-    }
-
-    public Habitacion getHabitacion() {
-        return habitacion;
-    }
-
-    public void setHabitacion(Habitacion habitacion) {
-        this.habitacion = habitacion;
-    }
-
-    public LocalDate getFechaInicio() {
-        return fechaInicio;
-    }
-
-    public void setFechaInicio(LocalDate fechaInicio) {
-        this.fechaInicio = fechaInicio;
-    }
-
-    public LocalDate getFechaFin() {
-        return fechaFin;
-    }
-
-    public void setFechaFin(LocalDate fechaFin) {
-        this.fechaFin = fechaFin;
-    }
-
-    public Estado getEstado() {
-        return estado;
-    }
-
-    public void setEstado(Estado estado) {
-        this.estado = estado;
-    }
-}
