--- src/main/java/org/example/enums/Cargo.java
@@ -1,9 +1,0 @@
-package org.example.enums;
-
-public enum Cargo {
-    ADMINISTRADOR,
-    CONSIERGE,
-    VALET,
-    LIMPIEZA,
-    MENTENIMIENTO
-}
