--- src/main/java/org/example/enums/Estado.java
@@ -1,7 +1,0 @@
-package org.example.enums;
-
-public enum Estado {
-    CONFIRMADO,
-    PENDIENTE,
-    CANCELADA;
-}
