--- src/main/java/org/example/enums/Segmento.java
@@ -1,7 +1,0 @@
-package org.example.enums;
-
-public enum Segmento {
-    BRONCE,
-    PLATA,
-    ORO
-}
