--- src/main/java/org/example/enums/TipoDeHabitacion.java
@@ -1,8 +1,0 @@
-package org.example.enums;
-
-public enum TipoDeHabitacion {
-    SIMPLE,
-    DOBLE,
-    SUITE,
-    DELUXE
-}
