--- src/main/java/org/example/excepciones/ObjectoNoEncontradoExcepcion.java
@@ -1,8 +1,0 @@
-package org.example.excepciones;
-
-public class ObjectoNoEncontradoExcepcion extends Exception{
-
-    public ObjectoNoEncontradoExcepcion(String mensaje){
-        super(mensaje);
-    }
-}
