--- src/main/java/org/example/excepciones/ObjetoYaExisteExcepcion.java
@@ -1,7 +1,0 @@
-package org.example.excepciones;
-
-public class ObjetoYaExisteExcepcion extends Exception{
-    public ObjetoYaExisteExcepcion(String mensaje){
-        super(mensaje);
-    }
-}
