--- src/main/java/org/example/gestor/GestorHotel.java
@@ -1,26 +1,0 @@
-package org.example.gestor;
-
-import org.example.gestor.impl.*;
-
-public class GestorHotel {
-
-    private final Hotel hotel;
-    private final GestorEmpleados gestorEmpleados;
-    private final GestorClientes gestorClientes;
-    private final GestorHabitaciones gestorHabitaciones;
-    private final GestorDeServicios gestorDeServicios;
-    private final GestorReservas gestorReservas;
-
-    public GestorHotel(Hotel hotel) {
-        this.hotel = hotel;
-        this.gestorClientes = new GestorClientes();
-        this.gestorHabitaciones = new GestorHabitaciones();
-        this.gestorDeServicios = new GestorDeServicios();
-        this.gestorReservas = new GestorReservas();
-        this.gestorEmpleados = new GestorEmpleados();
-    }
-
-    public GestorDeServicios getGestorDeServicios() {
-        return gestorDeServicios;
-    }
-}
