--- src/main/java/org/example/menues/HotelDelSurLanding.form
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="org.example.menues.HotelDelSurLanding">
-  <grid id="27dc6" binding="pane1" layout-manager="GridLayoutManager" row-count="5" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="149" y="87" width="500" height="186"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="cdd1a" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="47" height="39"/>
-          </grid>
-        </constraints>
-        <properties>
-          <text value="Usuario"/>
-        </properties>
-      </component>
-      <component id="cb7f9" class="javax.swing.JFormattedTextField" binding="formattedTextField1" default-binding="true">
-        <constraints>
-          <grid row="1" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="39"/>
-          </grid>
-        </constraints>
-        <properties>
-          <columns value="10"/>
-          <text value=""/>
-        </properties>
-      </component>
-      <component id="9764f" class="javax.swing.JButton" binding="salirButton" default-binding="true">
-        <constraints>
-          <grid row="4" column="2" row-span="1" col-span="1" vsize-policy="4" hsize-policy="3" anchor="2" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Salir"/>
-        </properties>
-      </component>
-      <component id="e1fb0" class="javax.swing.JButton" binding="submitLoginButton">
-        <constraints>
-          <grid row="3" column="2" row-span="1" col-span="1" vsize-policy="4" hsize-policy="3" anchor="2" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="84" height="35"/>
-          </grid>
-        </constraints>
-        <properties>
-          <horizontalAlignment value="0"/>
-          <text value="Ingresar"/>
-          <verticalAlignment value="0"/>
-        </properties>
-      </component>
-      <component id="b5529" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Contraseña"/>
-        </properties>
-      </component>
-      <component id="44e25" class="javax.swing.JPasswordField" binding="passwordField1" default-binding="true">
-        <constraints>
-          <grid row="2" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <columns value="20"/>
-          <text value=""/>
-        </properties>
-      </component>
-      <component id="11cb2" class="javax.swing.JLabel" binding="bienvenida">
-        <constraints>
-          <grid row="0" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="JetBrains Mono"/>
-          <iconTextGap value="4"/>
-          <text value="Bienvenido"/>
-          <toolTipText value=""/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
