+++ src/main/java/org/example/menues/acciones/AccionBotonListaTareas.java
@@ -1,0 +1,12 @@
+package org.example.menues.acciones;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class AccionBotonListaTareas implements ActionListener {
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+
+    }
+}
