+++ src/main/java/org/example/menues/enums/Tarea.java
@@ -1,0 +1,9 @@
+package org.example.menues.enums;
+
+public enum Tarea {
+    CLIENTES,
+    EMPLEADOS,
+    RESERVAS,
+    HABITACIONES,
+    ACTIVIDADES
+}
