+++ src/main/java/org/example/sistema/entidades/persona/Cliente.java
@@ -1,0 +1,25 @@
+package org.example.sistema.entidades.persona;
+
+import org.example.sistema.enums.Segmento;
+
+public class Cliente extends Persona{
+
+    private Segmento segmento;
+
+    public Cliente(String nombre, String apellido, String dni, Segmento segmento) {
+        super(nombre, apellido, dni);
+        this.segmento = segmento;
+    }
+
+    public Segmento getSegmento() {
+        return segmento;
+    }
+
+    public void setSegmento(Segmento segmento) {
+        this.segmento = segmento;
+    }
+
+
+
+
+}
