+++ src/main/java/org/example/sistema/entidades/persona/Persona.java
@@ -1,0 +1,51 @@
+package org.example.sistema.entidades.persona;
+
+import java.util.Objects;
+
+public abstract class Persona {
+    // private final String id;
+    private final String nombre;
+    private final String apellido;
+    private final String dni;
+
+    private boolean deleted = false;
+
+    public Persona(String nombre, String apellido, String dni) {
+        this.nombre = nombre;
+        this.apellido = apellido;
+        this.dni = dni;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public String getDni() {
+        return dni;
+    }
+
+    public boolean isDeleted() {
+        return deleted;
+    }
+
+    public void setDeleted(boolean deleted) {
+        this.deleted = deleted;
+    }
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Persona persona)) return false;
+        return Objects.equals(dni, persona.dni);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(dni);
+    }
+}
