+++ src/main/java/org/example/sistema/enums/Cargo.java
@@ -1,0 +1,9 @@
+package org.example.sistema.enums;
+
+public enum Cargo {
+    ADMINISTRADOR,
+    CONSIERGE,
+    VALET,
+    LIMPIEZA,
+    MENTENIMIENTO
+}
