+++ src/main/java/org/example/sistema/enums/Estado.java
@@ -1,0 +1,7 @@
+package org.example.sistema.enums;
+
+public enum Estado {
+    CONFIRMADO,
+    PENDIENTE,
+    CANCELADA;
+}
