+++ src/main/java/org/example/sistema/enums/Segmento.java
@@ -1,0 +1,7 @@
+package org.example.sistema.enums;
+
+public enum Segmento {
+    BRONCE,
+    PLATA,
+    ORO
+}
