+++ src/main/java/org/example/sistema/enums/TipoDeHabitacion.java
@@ -1,0 +1,8 @@
+package org.example.sistema.enums;
+
+public enum TipoDeHabitacion {
+    SIMPLE,
+    DOBLE,
+    SUITE,
+    DELUXE
+}
