+++ src/main/java/org/example/sistema/excepciones/ObjectoNoEncontradoExcepcion.java
@@ -1,0 +1,11 @@
+package org.example.sistema.excepciones;
+
+public class ObjectoNoEncontradoExcepcion extends Exception{
+
+    private final String parametro;
+
+    public ObjectoNoEncontradoExcepcion(String parametro){
+        super();
+        this.parametro = parametro;
+    }
+}
