+++ src/main/java/org/example/sistema/excepciones/ObjetoYaExisteExcepcion.java
@@ -1,0 +1,7 @@
+package org.example.sistema.excepciones;
+
+public class ObjetoYaExisteExcepcion extends Exception{
+    public ObjetoYaExisteExcepcion(String mensaje){
+        super(mensaje);
+    }
+}
